Issue #30, Enhanced cert handling

+++ pkg/keysupport/key_handler.go
@@ -22,18 +22,21 @@ import (
 	"bytes"
 	"crypto/rand"
 	"crypto/rsa"
+	"crypto/tls"
 	"crypto/x509"
 	"crypto/x509/pkix"
 	"encoding/pem"
 	"errors"
 	"fmt"
-	"log"
 	"math/big"
 	"net"
+	"net/http"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
+
+	log "golang.org/x/exp/slog"
 )
 
 const (
@@ -41,20 +44,24 @@ const (
 	EnvCertCountry   string = "HEXA_CERT_COUNTRY"
 	EnvCertProv      string = "HEXA_CERT_PROV"
 	EnvCertLocality  string = "HEXA_CERT_LOCALITY"
-	EnvCertCaKey     string = "HEXA_CA_KEYFILE"     // The location of a private key used to generate server keys
+	EnvCertCaPrivKey string = "HEXA_CA_KEYFILE" // The location of a private key used to generate server keys
+	EnvCertCaPubKey  string = "HEXA_CA_CERT"
 	EnvCertDirectory string = "HEXA_CERT_DIRECTORY" // The location where keys are stored.
-	EnvServerCert    string = "SERVER_CERT"
-	EnvServerKey     string = "SERVER_KEY_PATH"
-	EnvServerDNS     string = "SERVER_DNS_NAME"
+	EnvServerCert    string = "HEXA_SERVER_CERT"
+	EnvServerKey     string = "HEXA_SERVER_KEY_PATH"
+	EnvServerDNS     string = "HEXA_SERVER_DNS_NAME"
 	EnvAutoCreate    string = "HEXA_AUTO_SELFSIGN"
 )
 
 type KeyConfig struct {
-	KeyFile        string // The file containing a PEM encoded PKCS1 private key
+	CaKeyFile      string // The file containing a PEM encoded PKCS1 private key
+	CaCertFile     string
+	CaPrivKey      *rsa.PrivateKey
 	CertDir        string // This is the directory where generated keys are output
 	PkixName       pkix.Name
 	ServerCertPath string
 	ServerKeyPath  string
+	CaConfig       *x509.Certificate
 }
 
 /*
@@ -82,35 +89,52 @@ func GetKeyConfig() KeyConfig {
 
 	certDir := os.Getenv(EnvCertDirectory)
 	if certDir == "" {
-
 		file := os.Getenv("HOME")
 		certDir = filepath.Join(file, "./.certs")
-		fmt.Println("Defaulting certificate directory to: " + certDir)
+		log.Info("Defaulting certificate directory to: " + certDir)
 	}
 
 	// Create the directory if it does not exist
 	err := os.Mkdir(certDir, 0755)
 	if !os.IsExist(err) {
-		panic(fmt.Sprintf("Was unable to open or create certificate directory(%s):%s", certDir, err))
+		msg := fmt.Sprintf("Was unable to open or create certificate directory(%s):%s", certDir, err)
+		log.Error(msg)
+		panic(msg)
 	}
 
-	caKey := os.Getenv(EnvCertCaKey)
-	if caKey == "" {
-		caKey = filepath.Join(certDir, "ca-key.pem")
+	caCertKey := os.Getenv(EnvCertCaPrivKey)
+	if caCertKey == "" {
+		caCertKey = filepath.Join(certDir, "ca-key.pem")
+	}
+
+	caCert := os.Getenv(EnvCertCaPubKey)
+	if caCert == "" {
+		caCert = filepath.Join(certDir, "ca-cert.pem")
 	}
 
 	serverKeyPath := os.Getenv(EnvServerKey)
 	if serverKeyPath == "" {
 		serverKeyPath = filepath.Join(certDir, "server-key.pem")
+	} else {
+		if filepath.Dir(serverKeyPath) == "." {
+			// if the server key path is just a file name, prefix it with certDir
+			serverKeyPath = filepath.Join(certDir, serverKeyPath)
+		}
 	}
 
 	serverCertPath := os.Getenv(EnvServerCert)
 	if serverCertPath == "" {
 		serverCertPath = filepath.Join(certDir, "server-cert.pem")
+	} else {
+		if filepath.Dir(serverCertPath) == "." {
+			// if the server cert path is just a file name, prefix it with certDir
+			serverCertPath = filepath.Join(certDir, serverCertPath)
+		}
 	}
 
 	return KeyConfig{
-		KeyFile:        caKey,
+		CaKeyFile:      caCertKey,
+		CaCertFile:     caCert,
 		CertDir:        certDir,
 		ServerKeyPath:  serverKeyPath,
 		ServerCertPath: serverCertPath,
@@ -131,28 +155,32 @@ func (config KeyConfig) CertDirExists() bool {
 	return dirStat.IsDir()
 }
 func (config KeyConfig) RootKeyExists() bool {
-	_, err := os.Stat(config.KeyFile)
+	_, err := os.Stat(config.CaKeyFile)
 	return err == nil
 }
 
-func (config KeyConfig) ServerKeyExists() bool {
-	_, err := os.Stat(config.ServerKeyPath)
+func (config KeyConfig) ServerCertExists() bool {
+	_, err := os.Stat(config.ServerCertPath)
 	return err == nil
 }
 
 /*
-CreateSelfSignedKeys creates a set of self signed keys and writes them out to the directory in KeyConfig.CertDir
+InitializeKeys creates a set of self-signed keys and writes them out to the directory in KeyConfig.CertDir
 This includes:  Certificate Authority Certificate and Key (ca-cert/ca-key), Server certificate (server-cert.pem) and key
 (server-key.pem), and a client certificate (client-cert.pem) and key (client-key.pem).
 */
-func (config KeyConfig) CreateSelfSignedKeys() (err error) {
-	auto := os.Getenv(EnvAutoCreate)
-	if auto != "" && strings.ToLower(auto[0:1]) == "f" {
-		log.Println("Auto self-sign create is disabled (HEXA_AUTO_SELFSIGN). Will not generate.")
-		return nil
+func (config KeyConfig) InitializeKeys() (err error) {
+	autoFlag := os.Getenv(EnvAutoCreate)
+	auto := true
+	if autoFlag != "" && strings.ToLower(autoFlag[0:1]) == "f" {
+		log.Warn("Auto self-sign create is disabled (HEXA_AUTO_SELFSIGN). Will not generate keys.")
+		auto = false
 	}
+
+	var caPrivKey *rsa.PrivateKey
+
 	// set up our CA certificate
-	ca := &x509.Certificate{
+	config.CaConfig = &x509.Certificate{
 		SerialNumber:          big.NewInt(2019),
 		Subject:               config.PkixName,
 		NotBefore:             time.Now(),
@@ -163,74 +191,69 @@ func (config KeyConfig) CreateSelfSignedKeys() (err error) {
 		BasicConstraintsValid: true,
 	}
 
-	var caPrivKey *rsa.PrivateKey
+	if config.RootKeyExists() {
 
-	keyFile := config.KeyFile
-	if keyFile != "" {
-		_, err := os.Stat(keyFile) // first check that it exists, otherwise fall through to create
-		if err == nil {
-			log.Println(fmt.Sprintf("Attempting to load existing CA Key from %s ...", keyFile))
-			keyBytes, err := os.ReadFile(keyFile)
+		log.Info(fmt.Sprintf("Loading existing CA Key from %s ...", config.CaKeyFile))
+		keyBytes, err := os.ReadFile(config.CaKeyFile)
+		if err != nil {
+			return err
+		}
+		if len(keyBytes) > 10 {
+			pemBlock, _ := pem.Decode(keyBytes)
+			if pemBlock == nil {
+				return errors.New("expecting file to contain a PEM key")
+			}
+			caBytes := pemBlock.Bytes
+			caPrivKey, err = x509.ParsePKCS1PrivateKey(caBytes)
+			config.CaPrivKey = caPrivKey
 			if err != nil {
 				return err
 			}
-			if len(keyBytes) > 10 {
-				pemBlock, _ := pem.Decode(keyBytes)
-				if pemBlock == nil {
-					return errors.New("expecting file to contain a PEM key")
-				}
-				caBytes := pemBlock.Bytes
-				caPrivKey, err = x509.ParsePKCS1PrivateKey(caBytes)
-				if err != nil {
-					return err
-				}
-			}
 		}
-	}
-	if caPrivKey == nil {
-		log.Println("Generating new CA key...")
+
+	} else if auto {
+
+		log.Info("Generating new CA key pair...")
 		// create our private and public key
 		caPrivKey, err = rsa.GenerateKey(rand.Reader, 4096)
 		if err != nil {
 			return err
 		}
+		config.CaPrivKey = caPrivKey
 
 		caPrivKeyPEM := new(bytes.Buffer)
 		_ = pem.Encode(caPrivKeyPEM, &pem.Block{
 			Type:  "RSA PRIVATE KEY",
 			Bytes: x509.MarshalPKCS1PrivateKey(caPrivKey),
 		})
-		log.Println("Writing out CA Key")
-		err = os.WriteFile(config.KeyFile, caPrivKeyPEM.Bytes(), 0644)
+		log.Debug("Writing out CA Key to: " + config.CaKeyFile)
+		err = os.WriteFile(config.CaKeyFile, caPrivKeyPEM.Bytes(), 0644)
 		if err != nil {
 			return err
 		}
 
-	}
+		caBytes, err := x509.CreateCertificate(rand.Reader, config.CaConfig, config.CaConfig, &caPrivKey.PublicKey, caPrivKey)
+		if err != nil {
+			return err
+		}
+		// pem encode
+		caPEM := new(bytes.Buffer)
+		_ = pem.Encode(caPEM, &pem.Block{
+			Type:  "CERTIFICATE",
+			Bytes: caBytes,
+		})
 
-	// create the CA Public key file (in case it does not exist)
-	caBytes, err := x509.CreateCertificate(rand.Reader, ca, ca, &caPrivKey.PublicKey, caPrivKey)
-	if err != nil {
-		return err
-	}
+		log.Debug("Writing out CA Public Cert to: " + config.CaCertFile)
 
-	// pem encode
-	caPEM := new(bytes.Buffer)
-	_ = pem.Encode(caPEM, &pem.Block{
-		Type:  "CERTIFICATE",
-		Bytes: caBytes,
-	})
-	log.Println("Writing out CA Cert")
-
-	err = os.WriteFile(filepath.Join(config.CertDir, "ca-cert.pem"), caPEM.Bytes(), 0644)
-	if err != nil {
-		return err
+		err = os.WriteFile(config.CaCertFile, caPEM.Bytes(), 0644)
+		if err != nil {
+			return err
+		}
 	}
 
 	// set up our server certificate
 
-	dnsNames := []string{}
-
+	var dnsNames []string
 	domainName := os.Getenv(EnvServerDNS)
 	if domainName == "" {
 		dnsNames = []string{"hexa_bundle-server", "hexa-opaBundle-server"}
@@ -238,30 +261,37 @@ func (config KeyConfig) CreateSelfSignedKeys() (err error) {
 		dnsNames = strings.Split(domainName, ",")
 	}
 
-	certPEM, certPrivKeyPEM, err := config.generateCert(
-		ca,
-		caPrivKey,
-		[]x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth, x509.ExtKeyUsageServerAuth},
-		dnsNames,
-	)
-	if err != nil {
-		return err
+	if !config.ServerCertExists() {
+		log.Info("Generating server key pair")
+		certPEM, certPrivKeyPEM, err := config.generateCert(
+			config.CaConfig,
+			caPrivKey,
+			[]x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth, x509.ExtKeyUsageServerAuth},
+			dnsNames,
+		)
+		if err != nil {
+			return err
+		}
+		log.Debug("Writing out Server Cert to: " + config.ServerCertPath)
+		err = os.WriteFile(config.ServerCertPath, certPEM, 0644)
+		if err != nil {
+			return err
+		}
+		log.Debug("Writing out Server Key to: " + config.ServerKeyPath)
+		err = os.WriteFile(config.ServerKeyPath, certPrivKeyPEM, 0644)
+		if err != nil {
+			return err
+		}
 	}
-	log.Println("Writing out Server Cert")
-	err = os.WriteFile(config.ServerCertPath, certPEM, 0644)
-	if err != nil {
-		return err
-	}
-	log.Println("Writing out Server Key")
-	err = os.WriteFile(config.ServerKeyPath, certPrivKeyPEM, 0644)
-	if err != nil {
-		return err
-	}
+	return
+}
 
+func (config KeyConfig) GenerateClientKeys(keyPath, certPath string) (err error) {
 	// set up our client certificate
+	log.Info("Generating client keys")
 	clientCertPEM, clientCertPrivKeyPEM, err := config.generateCert(
-		ca,
-		caPrivKey,
+		config.CaConfig,
+		config.CaPrivKey,
 		[]x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth},
 		[]string{},
 	)
@@ -269,18 +299,24 @@ func (config KeyConfig) CreateSelfSignedKeys() (err error) {
 		return err
 	}
 
-	log.Println("Writing out Client Cert")
-	err = os.WriteFile(filepath.Join(config.CertDir, "client-cert.pem"), clientCertPEM, 0644)
+	if certPath == "" {
+		certPath = filepath.Join(config.CertDir, "client-cert.pem")
+	}
+	if keyPath == "" {
+		keyPath = filepath.Join(config.CertDir, "client-key.pem")
+	}
+	log.Info("Writing out Client Cert to: " + certPath)
+	err = os.WriteFile(certPath, clientCertPEM, 0644)
 	if err != nil {
 		return err
 	}
-	log.Println("Writing out Client Key")
-	err = os.WriteFile(filepath.Join(config.CertDir, "client-key.pem"), clientCertPrivKeyPEM, 0644)
+	log.Info("Writing out Client Key to: " + keyPath)
+	err = os.WriteFile(keyPath, clientCertPrivKeyPEM, 0644)
 	if err != nil {
 		return err
 	}
 
-	return
+	return nil
 }
 
 func (config KeyConfig) generateCert(
@@ -322,4 +358,34 @@ func (config KeyConfig) generateCert(
 		Bytes: x509.MarshalPKCS1PrivateKey(certPrivKey),
 	})
 	return certPEM.Bytes(), certPrivKeyPEM.Bytes(), nil
+}
+
+// CheckCaInstalled will check if a CA certificate has been installed in the http.Client or if nil, the system cert pool
+func CheckCaInstalled(client *http.Client) {
+	// Note; this is not tested because we don't want to install temporary test certs.
+	caCertPath := os.Getenv(EnvCertCaPubKey)
+
+	if caCertPath != "" {
+		caCertPem, err := os.ReadFile(caCertPath)
+		if err != nil {
+			log.Warn("Error reading CA certificate: " + err.Error())
+		}
+		var caPool *x509.CertPool
+		if client != nil {
+			log.Debug("Installing CA certificate into HTTP client", "file", caCertPath)
+			caPool = x509.NewCertPool()
+			t := &http.Transport{
+				TLSClientConfig: &tls.Config{RootCAs: caPool},
+			}
+			client.Transport = t
+		} else {
+			log.Debug("Installing CA certificate into system cert pool", "file", caCertPath)
+			caPool, _ = x509.SystemCertPool()
+		}
+		ok := caPool.AppendCertsFromPEM(caCertPem)
+		if !ok {
+			log.Error("Error loading CA PEM")
+		}
+
+	}
 }
